Moved Discover subcomponents into own folders
