+++ src/pages/Discover/Places/PlaceDetails.js
@@ -2,10 +2,10 @@ import React, { Component, Fragment } from "react";
 import { connect } from "react-redux";
 import { Tab, Card, Button, Item, Rating, Label, List, Divider, Grid } from "semantic-ui-react";
 
-import BasicMap from "../../components/GoogleMap/BasicMap";
-import Marker from "../../components/GoogleMap/Marker";
+import BasicMap from "../../../components/GoogleMap/BasicMap";
+import Marker from "../../../components/GoogleMap/Marker";
 
-import { updateActiveDiscover } from "../../actions";
+import { updateActiveDiscover } from "../../../actions";
 
 class PlaceDetails extends Component {
